CEF patchery: Giving up, rewriting from scratch.
This commit is just to record the current state, just in case someone _might_
need this. This stuff here also contains serveral CEF-unrelated stuff right now,
so don't wonder if it looks messy ;-)

This commit/branch might vanish or get rebased, so don't rely on this if you
really are that insane to start working based on this.